typescript: Move TS files into JS' directory.
