+++ frontend_tests/zjsunit/index.js
@@ -1,7 +1,7 @@
 require('ts-node').register({
-    project: 'static/ts/tsconfig.json',
+    project: 'static/js/tsconfig.json',
     compilerOptions: {
-        typeRoots: ["node_modules/@types", "../../static/ts/js_typings"],
+        typeRoots: ["node_modules/@types", "../../static/js/js_typings"],
         // We don't have webpack to handle es6 modules here so directly
         // transpile to CommonJS format.
         module: "commonjs",
@@ -27,7 +27,7 @@ global.window = _.extend({}, windowObj, {
     },
 });
 
-global.Dict = require('ts/dict.ts').Dict;
+global.Dict = require('js/dict').Dict;
 
 // Create a helper function to avoid sneaky delays in tests.
 function immediate(f) {
