+++ static/js/pm_conversations.js
@@ -1,4 +1,4 @@
-var Dict = require('../ts/dict').Dict;
+var Dict = require('./dict').Dict;
 
 var pm_conversations = (function () {
