+++ static/js/tsconfig.json
