+++ static/js/tslint.json
