+++ static/js/typing_data.js
@@ -1,4 +1,4 @@
-var Dict = require('../ts/dict').Dict;
+var Dict = require('./dict').Dict;
 
 var typing_data = (function () {
 var exports = {};
